feat: add handle floating-point numbers

+++ src/carbon-dating.js
@@ -4,7 +4,7 @@ const MODERN_ACTIVITY= 15;
 const HALF_LIFE_PERIOD= 5730;
 
 module.exports = function dateSample(sampleActivity) {
-  const sampleActivityNormalized = parseInt(sampleActivity);
+  const sampleActivityNormalized = parseFloat(sampleActivity);
 
   if (
     typeof sampleActivity !== 'string'||
